+++ src/main/java/kg/job/jobsearch/dao/mapper/ContactTypeMapper.java
@@ -11,7 +11,7 @@ public class ContactTypeMapper implements RowMapper<ContactType> {
     @Override
     public ContactType mapRow(ResultSet rs, int rowNum) throws SQLException {
         ContactType contactType = new ContactType();
-        contactType.setId(rs.getInt("id"));
+        contactType.setId(rs.getLong("id"));
         contactType.setType(rs.getString("type"));
         return contactType;
     }
