+++ src/main/java/kg/job/jobsearch/dao/mapper/ContactsInfoMapper.java
@@ -11,9 +11,9 @@ public class ContactsInfoMapper implements RowMapper<ContactsInfo> {
     @Override
     public ContactsInfo mapRow(ResultSet rs, int rowNum) throws SQLException {
         ContactsInfo contactsInfo = new ContactsInfo();
-        contactsInfo.setId(rs.getInt("id"));
-        contactsInfo.setType_id(rs.getInt("type_id"));
-        contactsInfo.setResume_id(rs.getInt("resume_id"));
+        contactsInfo.setId(rs.getLong("id"));
+        contactsInfo.setType_id(rs.getLong("type_id"));
+        contactsInfo.setResume_id(rs.getLong("resume_id"));
         contactsInfo.setContactValue(rs.getString("contact_value"));
         return contactsInfo;
     }
