+++ src/main/java/kg/job/jobsearch/model/EducationInfo.java
@@ -8,8 +8,8 @@ import java.time.LocalDate;
 @Getter
 @Setter
 public class EducationInfo {
-    private Integer id;
-    private Integer resumeId;
+    private Long id;
+    private Long resumeId;
     private String institution;
     private String program;
     private LocalDate startDate;
