+++ src/main/java/kg/job/jobsearch/model/Resume.java
@@ -9,10 +9,10 @@ import java.time.LocalDateTime;
 @Getter
 @Setter
 public class Resume {
-    private Integer id;
-    private Integer applicantId;
+    private Long id;
+    private Long applicantId;
     private String name;
-    private Integer categoryId;
+    private Long categoryId;
     private BigDecimal salary;
     private Boolean isActive;
     private LocalDateTime createdDate;
